+++ src/constants/categoryIdMap.js
@@ -30,11 +30,3 @@ export const CATEGORY_ID_MAP = {
   KIDS: {},
   EVENT: {},
 };
-
-export function getCategoryId(primary, item) {
-  if (!primary || !item) return null;
-  const P = String(primary).toUpperCase();
-  const bucket = CATEGORY_ID_MAP[P];
-  if (!bucket) return null;
-  return bucket[String(item)] ?? null;
-}
